WEB_APP/requirements.txt
flask
dpkt
pendulum
fuzzywuzzy
python-Levenshtein
yara-python
geoip2
pandas
fastapi
uvicorn